Naive fix for paper-input in Polymer 1.0 which fires bind-value-changed
Summary:
This is not a real fix, just a draft of how things can be done.
We need to extend the mapping structure to include information
about which events/properties to bind for each of the components,
as the polymer implementation is not yet consistent about the
two-way binding as it is described in the 1.0 docs.

Test Plan:
Change bower.json in your web project to use this version
for ng-polymer-elements:
https://github.com/m-creations/ng-polymer-elements.git#master

Reviewers: #m-creations

Differential Revision: https://phab.m-creations.net/D155